Fix possible compilation failure introduced in #122
